add weapon helper functions and constants for projectile mechanics
