+++ src/FPSGame.tsx
@@ -1,2 +1,3 @@
 import './HealthSystem'
 import './weapons/WeaponSystem'
+import './weapons/functions/weaponHelper'
